ex_venture v0.15.0 Data.Item.Compiled

An item is compiled after the item tags are rolled together and merged with the base item’s stats.

Types

t()

t() :: %Data.Item.Compiled{cost: term, description: term, drop_rate: term, effects: term, id: term, keywords: term, level: term, name: term, stats: term, type: term}

Functions

compile(item)

The item’s item tags should be preloaded

Repo.preload(item, [item_taggings: [:item_tag]])

merge_effects(compiled_item, map)

merge_effects(compiled_item :: t, item :: Data.Item.t) :: t

Concatenate effects of the item and all of its tags

merge_stats(compiled_item, map)

merge_stats(compiled_item :: t, item :: Data.Item.t) :: t

Merge stats together

scale_for_level(level, value)

Scales a value for a level. Every 10 levels the value will double

iex> Data.Item.Compiled.scale_for_level(1, 5)
5
iex> Data.Item.Compiled.scale_for_level(11, 5)
10
iex> Data.Item.Compiled.scale_for_level(11, 10)
20