ex_venture v0.15.0 Data.Stats

Item statistics

Types

armor()

armor() :: %{slot: :atom}

character()

character() :: %{health: integer, max_health: integer, skill_points: integer, max_skill_points: integer, move_points: integer, max_move_points: integer, strength: integer, intelligence: integer, dexterity: integer}

weapon()

weapon() :: %{}

Functions

default(stats)

default(Stats.t) :: Stats.t

Set defaults for new statistics

A “migration” of stats to ensure new ones are always available. They should be saved back in after the user loads their account.

fields()

fields() :: [atom]

Fields in the statistics map

slots()

slots() :: [atom]

Slots on a character

valid?(type, stats)

valid?(type :: String.t, stats :: Stats.t) :: boolean

Validate an item stats based on type

iex> Data.Stats.valid?("basic", %{})
true
iex> Data.Stats.valid?("basic", %{slot: :chest})
false

valid_armor?(stats)

valid_armor?(stats :: Stats.armor) :: boolean

Validate an armor item

iex> Data.Stats.valid_armor?(%{slot: :chest, armor: 10})
true

iex> Data.Stats.valid_armor?(%{slot: :chest, armor: :none})
false

iex> Data.Stats.valid_armor?(%{slot: :eye, armor: 10})
false

iex> Data.Stats.valid_armor?(%{})
false

valid_character?(stats)

valid_character?(stats :: Stats.character) :: boolean

Validate a character’s stats

iex> Data.Stats.valid_character?(%{health: 50, strength: 10})
false

iex> Data.Stats.valid_character?(%{})
false

valid_damage?(stats)

valid_damage?(stats :: Stats.t) :: boolean

Validate if the damage is right

iex> Data.Stats.valid_damage?(%{damage_type: :slashing, damage: 10})
true
iex> Data.Stats.valid_damage?(%{damage_type: :slashing, damage: nil})
false
iex> Data.Stats.valid_damage?(%{damage_type: :finger})
false

valid_slot?(stats)

valid_slot?(stats :: Stats.t) :: boolean

Validate if the slot is right

iex> Data.Stats.valid_slot?(%{slot: :chest})
true
iex> Data.Stats.valid_slot?(%{slot: :eye})
false

valid_weapon?(stats)

valid_weapon?(stats :: Stats.weapon) :: boolean

Validate a weapon item

iex> Data.Stats.valid_weapon?(%{})
true

iex> Data.Stats.valid_weapon?(%{anything: true})
false