ex_venture v0.15.0 Game.Character

Character GenServer client

A character is a player (session genserver) or an NPC (genserver). They should handle the following casts:

{:targeted, player}

{:remove_target, player}

{:apply_effects, effects, player}

{:died, player}

Functions

apply_effects(target, effects, from, description)

apply_effects(target :: tuple, effects :: [Effect.t], from :: {atom, map}, description :: String.t) :: :ok

Apply effects on the target

being_targeted(target, player)

being_targeted(target :: tuple, player :: {atom, Data.User.t}) :: :ok

Let the target know they are being targeted

died(target, who)

died(who :: {atom, map}, to :: {atom, map}) :: :ok

Let the character targeting you know you died

PC targets NPC, NPC dies, NPC let’s the PC know it died. Should clear the target on the PC.

remove_target(target, player)

remove_target(target :: tuple, player :: {atom, Data.User.t}) :: :ok

When a player stops targetting a character, let them know

who(arg)

Converts a tuple with a struct to a tuple with an id