ex_venture v0.15.0 Game.Command behaviour

Parses and runs commands from players

Types

t()

t() :: {module :: atom, args :: []}

A tuple with the first element being the command to run

Example:

{Game.Command.Run, "Hi there"}

Functions

commands()

commands() :: [atom]

Get all commands that use Command

parse(command, user)

parse(command :: String.t, user :: Data.User.t) :: t

Parse a string to turn into a command tuple

parse_command(module, command)

parse_command(module :: atom, command :: String.t) :: [String.t]

Parse a command

Uses the module’s commands and aliases to find the arguments

iex> Game.Command.parse_command(Game.Command.Who, "who")
{}

iex> Game.Command.parse_command(Game.Command.Say, "say hi")
{"hi"}

run(command, session, state)

Callbacks

run(args, session, state)

run(args :: list, session :: pid, state :: map) ::
  :ok |
  {:update, state :: map}

Run a command

Returns :ok or {:update, new_state} and the Session server will accept the new state.