ex_venture v0.15.0 Game.Command.Wield

The “wield” command

Functions

parse(command)

parse(command :: String.t) :: []

Parse the command to determine wield or unwield

iex> Game.Command.Wield.parse("wield right sword")
{:wield, "right sword"}

iex> Game.Command.Wield.parse("unwield right sword")
{:unwield, "right sword"}

iex> Game.Command.Wield.parse("unweld right sword")
{:error, :bad_parse, "unweld right sword"}

pick_hand(item)

pick_hand(item_string :: String.t) ::
  {:right, String.t} |
  {:left, String.t}

Determine which hand is being used from the item string

Examples:

iex> Game.Command.Wield.pick_hand("sword")
{:right, "sword"}

iex> Game.Command.Wield.pick_hand("right sword")
{:right, "sword"}

iex> Game.Command.Wield.pick_hand("left sword")
{:left, "sword"}

run(command, session, state)

run(args :: {atom, String.t}, session :: Game.Session.t, state :: map) :: :ok

Put an item in your hands

unwield(hand, wielding, item_ids)

unwield(hand :: atom, wielding :: map, item_ids :: [integer]) :: {wielding :: map, inventory :: [integer]}

Remove an item from your hand and place back into the inventory

Examples:

iex> Game.Command.Wield.unwield(:right, %{right: 1}, [])
{%{}, [1]}
iex> Game.Command.Wield.unwield(:right, %{right: 1, left: 2}, [3])
{%{left: 2}, [1, 3]}

iex> Game.Command.Wield.unwield(:left, %{left: 1}, [])
{%{}, [1]}
iex> Game.Command.Wield.unwield(:left, %{right: 1, left: 2}, [3])
{%{right: 1}, [2, 3]}

iex> Game.Command.Wield.unwield(:right, %{}, [])
{%{}, []}
iex> Game.Command.Wield.unwield(:right, %{left: 1}, [])
{%{left: 1}, []}

iex> Game.Command.Wield.unwield(:left, %{}, [])
{%{}, []}
iex> Game.Command.Wield.unwield(:left, %{right: 1}, [])
{%{right: 1}, []}

iex> Game.Command.Wield.unwield(:right, nil, [])
{%{}, []}
iex> Game.Command.Wield.unwield(:left, nil, [])
{%{}, []}