ex_venture v0.15.0 Game.Door

Door tracker, know if the door is open or closed

Functions

clear()

For testing only: clear the EST table

closed?(exit_id)

closed?(exit_id :: integer) :: boolean

Check if a door is closed

get(exit_id)

get(exit_id :: integer) :: String.t

Get the state of a door

load(exit_id)

load(exit_id :: integer) :: :ok

Load a new door into the ETS table

maybe_load(room_exit)

maybe_load(room_exit :: Data.Exit.t) :: :ok

Maybe load a door, only if the exit has a door

open?(exit_id)

open?(exit_id :: integer) :: boolean

Check if a door is open

remove(map)

set(exit_id, state)

set(exit_id :: integer, state :: String.t) :: :ok

Set the state of a door, state must be open or closed