ex_venture v0.15.0 Game.Effect

Calculate and apply effects from skills/items

Functions

apply(effects, stats)

apply(effects :: [Data.Effect.t], stats :: Data.Stats.t) :: Data.Stats.t

Apply effects to stats

iex> effects = [%{kind: "damage", type: :slashing, amount: 10}]
iex> Game.Effect.apply(effects, %{health: 25})
%{health: 15}

apply_effect(effect, stats)

apply_effect(effect :: Data.Effect.t, stats :: Data.Stats.t) :: Data.Stats.t

Apply an effect to stats

iex> effect = %{kind: "damage", type: :slashing, amount: 10}
iex> Game.Effect.apply_effect(effect, %{health: 25})
%{health: 15}

calculate(stats, effects)

calculate(stats :: Data.Stats.t, effects :: [Data.Effect.t]) :: [map]

Calculate effects based on the user

Filters out stat boosting effects, then deals with damage

iex> Game.Effect.calculate(%{}, [])
[]

iex> Game.Effect.calculate(%{strength: 10}, [%{kind: "damage", type: :slashing, amount: 10}])
[%{kind: "damage", amount: 15, type: :slashing}]

calculate_damage(effect, stats)

calculate_damage(effect :: Data.Effect.t, stats :: Data.Stats.t) :: map

Calculate damage

Physical:

iex> effect = %{kind: "damage", amount: 10, type: :slashing}
iex> Game.Effect.calculate_damage(effect, %{strength: 10})
%{kind: "damage", amount: 15, type: :slashing}

Magical:

iex> effect = %{kind: "damage", amount: 10, type: :arcane}
iex> Game.Effect.calculate_damage(effect, %{intelligence: 10})
%{kind: "damage", amount: 15, type: :arcane}

calculate_damage_type(effect, damages)

calculate_damage_type(effect :: Data.Effect.t, stats :: Data.Stats.t) :: map

Calculate damage type effects

Damage:

iex> effect = %{kind: "damage/type", types: [:slashing]}
iex> damage = %{kind: "damage", amount: 10, type: :bludgeoning}
iex> Game.Effect.calculate_damage_type(effect, [damage])
[%{kind: "damage", amount: 5, type: :bludgeoning}]

calculate_stats(stats, effects)

calculate_stats(stats :: Data.Stats.t, effects :: [Data.Effect.t]) :: Data.Stats.t

Calculate stats and return any effects that were not processed

iex> stats = %{strength: 10} iex> effects = [%{kind: “stats”, field: :strength, amount: 10}, %{kind: “damage”}] iex> Game.Effect.calculate_stats(stats, effects) {%{strength: 20}, [%{kind: “damage”}]}

process_stats(effect, stats)

process_stats(effect :: Data.Effect.t, stats :: Data.Stats.t) :: Data.Stats.t

Process stats effects

iex> Game.Effect.process_stats(%{field: :strength, amount: 10}, %{strength: 10})
%{strength: 20}