ex_venture v0.15.0 Game.Format

Format data into strings to send to the connected player

Functions

channel_say(channel, sender, message)

channel_say(channel :: String.t, sender :: {atom, map}, message :: String.t) :: String.t

Format a channel message

Example:

iex> Game.Format.channel_say("global", {:npc, %{name: "NPC"}}, "Hello")
~s({red}[global]{/red} {yellow}NPC{/yellow} says, {green}"Hello"{/green})

currency(map)

currency(Data.Save.t | Data.Room.t) :: String.t

Format currency

dropped(who, item)

dropped(who :: {tuple, map}, item :: Data.Item.t) :: String.t

An item was dropped message

iex> Game.Format.dropped({:npc, %{name: "NPC"}}, %{name: "Sword"})
"{yellow}NPC{/yellow} dropped a Sword."

iex> Game.Format.dropped({:user, %{name: "Player"}}, %{name: "Sword"})
"{blue}Player{/blue} dropped a Sword."

iex> Game.Format.dropped({:user, %{name: "Player"}}, {:currency, 100})
"{blue}Player{/blue} dropped 100 gold."

effects(list)

Format effects for display.

emote(arg, emote)

emote(sender :: map, message :: String.t) :: String.t

Format an emote message

Example:

iex> Game.Format.emote({:npc, %{name: "NPC"}}, "does something")
~s[{yellow}NPC{/yellow} {green}does something{/green}]

iex> Game.Format.emote({:user, %{name: "Player"}}, "does something")
~s[{blue}Player{/blue} {green}does something{/green}]

equipment(wearing, wielding)

equipment(wearing :: map, wielding :: map) :: String.t

Format your equipment

Example:

iex> wearing = %{chest: %{name: "Leather Armor"}}
iex> wielding = %{right: %{name: "Short Sword"}, left: %{name: "Shield"}}
iex> Game.Format.equipment(wearing, wielding)
"You are wearing:\n  - {cyan}Leather Armor{/cyan} on your chest\nYou are wielding:\n  - a {cyan}Shield{/cyan} in your left hand\n  - a {cyan}Short Sword{/cyan} in your right hand"

info(user)

info(user :: Data.User.t) :: String.t

Format your info sheet

inventory(currency, wearing, wielding, items)

inventory(currency :: integer, wearing :: map, wielding :: map, items :: [Data.Item.t]) :: String.t

Format your inventory

item(item)

item(item :: Data.Item.t) :: String.t

Display an item

Example:

iex> string = Game.Format.item(%{name: "Short Sword", description: "A simple blade"})
iex> Regex.match?(~r(Short Sword), string)
true

item_stats(item)

item_stats(Data.Item.t) :: String.t

Format an items stats

iex> Game.Format.item_stats(%{type: "armor", stats: %{slot: :chest}})
"Slot: chest"

iex> Game.Format.item_stats(%{type: "basic"})
""

items(room)

list_shop(shop, items)

name(who)

npcs(arg1)

npcs(room :: Game.Room.t) :: String.t

Format NPC text for who is in the room

Example:

iex> Game.Format.npcs(%{npcs: [%{name: "Mordred"}, %{name: "Arthur"}]})
"{yellow}Mordred{/yellow} is here. {yellow}Arthur{/yellow} is here."

peak_room(room, direction)

peak_room(room :: Game.Room.t, direction :: String.t) :: String.t

Peak at a room from the room you’re in

Example:

iex> Game.Format.peak_room(%{name: “Hallway”}, “north”) “{green}Hallway{/green} is north.”

play_time(seconds)

play_time(seconds :: integer) :: String.t

Format number of seconds online into a human readable string

iex> Game.Format.play_time(125)
"00h 02m 05s"

iex> Game.Format.play_time(600)
"00h 10m 00s"

iex> Game.Format.play_time(3670)
"01h 01m 10s"

iex> Game.Format.play_time(36700)
"10h 11m 40s"

players(arg1)

players(room :: Game.Room.t) :: String.t

Format Player text for who is in the room

Example:

iex> Game.Format.players(%{players: [%{name: "Mordred"}, %{name: "Arthur"}]})
"{blue}Mordred{/blue} is here. {blue}Arthur{/blue} is here."

prompt(user, save)

prompt(user :: Data.User.t, save :: Data.Save.t) :: String.t

Format the user’s prompt

Example:

iex> class = %{points_abbreviation: "SP"}
iex> stats = %{health: 50, max_health: 75, skill_points: 9, max_skill_points: 10, move_points: 4, max_move_points: 10}
iex> Game.Format.prompt(%{name: "user", class: class}, %{experience_points: 1010, stats: stats})
"[50/75hp 9/10sp 4/10mv 10xp] > "

room(room, map)

room(room :: Game.Room.t, map :: String.t) :: String.t

Format full text for a room

say(arg, message)

say(sender :: map, message :: String.t) :: String.t

Format a say message

Example:

iex> Game.Format.say({:npc, %{name: "NPC"}}, "Hello")
~s[{yellow}NPC{/yellow} says, {green}"Hello"{/green}]

iex> Game.Format.say({:user, %{name: "Player"}}, "Hello")
~s[{blue}Player{/blue} says, {green}"Hello"{/green}]

shops(room, opts \\ [])

Format Shop text for shops in the room

Example:

iex> Game.Format.shops(%{shops: [%{name: "Hole in the Wall"}]})
"Shops: {magenta}Hole in the Wall{/magenta}"

iex> Game.Format.shops(%{shops: [%{name: "Hole in the Wall"}]}, label: false)
"  - {magenta}Hole in the Wall{/magenta}"

skill(skill, sp_name)

skill(skill :: Data.Skill.t, sp_name :: String.t) :: String.t

Format a skill

iex> skill = %{level: 1, name: "Slash", points: 2, command: "slash", description: "Fight your foe"}
iex> Game.Format.skill(skill, "PP")
"1 - Slash (slash) - 2PP - Fight your foe"

skill_usee(skill, skill_effects, skill_user)

Format a skill, from the perspective of a usee

iex> Game.Format.skill_usee(%{usee_text: "Slash away"}, [], {:npc, %{name: "Bandit"}})
"Slash away"

iex> effects = [%{kind: "damage", type: :slashing, amount: 10}]
iex> Game.Format.skill_usee(%{usee_text: "You were slashed at by {user}"}, effects, {:npc, %{name: "Bandit"}})
"You were slashed at by {yellow}Bandit{/yellow}\n10 slashing damage is dealt."

skill_user(skill, effects, target)

Format a skill, from perspective of the user

iex> Game.Format.skill_user(%{user_text: "Slash away"}, [], {:npc, %{name: "Bandit"}})
"Slash away"

iex> effects = [%{kind: "damage", type: :slashing, amount: 10}]
iex> Game.Format.skill_user(%{user_text: "You slash away at {target}"}, effects, {:npc, %{name: "Bandit"}})
"You slash away at {yellow}Bandit{/yellow}\n10 slashing damage is dealt."

skills(class)

skills(class :: Data.Class.t) :: String.t

Format skills

iex> skills = [%{level: 1, name: "Slash", points: 2, command: "slash", description: "Fight your foe"}]
iex> Game.Format.skills(%{name: "Fighter", points_abbreviation: "PP", skills: skills})
"Fighter\n\nLevel - Name - Points - Description\n1 - Slash (slash) - 2PP - Fight your foe\n"

target_name(arg)

target_name({atom, map}) :: String.t

Format a target name, blue for user, yellow for npc

iex> Game.Format.target_name({:user, %{name: “Player”}}) “{blue}Player{/blue}”

iex> Game.Format.target_name({:npc, %{name: “Bandit”}}) “{yellow}Bandit{/yellow}”

tell(map, message)

Format a tell message

iex> Game.Format.tell(%{name: "Player"}, "secret message")
~s[{blue}Player{/blue} tells you, {green}"secret message"{/green}]

underline(string)

Create an ‘underline’

Example:

iex> Game.Format.underline("Room Name")
"-------------"

who_is_here(room)

Format full text for who is in the room

Example:

iex> Game.Format.who_is_here(%{players: [%{name: "Mordred"}], npcs: [%{name: "Arthur"}]})
"{blue}Mordred{/blue} is here. {yellow}Arthur{/yellow} is here."

wrap(string)

wrap(string :: String.t) :: String.t

Wraps lines of text