ex_venture v0.15.0 Game.Map

Map out a zone

Functions

display_map(zone, arg, opts \\ [])

display_map(zone :: Zone.t, {x :: integer, y :: integer, layer :: integer}, opts :: Keyword.t) :: [String.t]

Generate a text view of the zone

display_room(arg, coords)

Determine what the room looks like with it’s walls

join_line(last_line, new_line)

Join a line together, zipping each point

join_point(point, point)

join_point(String.t, String.t) :: String.t

Join a point in the map together

iex> Game.Map.join_point(" ", " ")
" "
iex> Game.Map.join_point("|", "+")
"+"

iex> Game.Map.join_point(" ", "+")
"+"
iex> Game.Map.join_point("+", " ")
"+"

iex> Game.Map.join_point(" ", "-")
"-"
iex> Game.Map.join_point("-", " ")
"-"

iex> Game.Map.join_point("|", " ")
"|"
iex> Game.Map.join_point("|", " ")
"|"

join_rooms(rows)

Join map together

The map is a list of rows of rooms, each room row is 3 lines

join_row(room, row)

Join an individual row of rooms together

join_rows(row, rows)

Join each row of rooms together

Each row of rooms are 3 lines long. The first line of the newly added row will be joined against the last row of the set.

layers_in_map(zone)

layers_in_map(zone :: Zone.t) :: [integer]

Determine which layers are in a map

map(zone, opts \\ [])

Generate a sorted list of lists for a zone map

mini_map(zone, arg2, bool)

mini_map(zone :: [], {integer, integer}, opts :: Keyword.t) :: []

Create a mini map of the whole zone. Restricts to with in 2 spaces of the player

room_color(room)

room_color(room :: Room.t) :: String.t

Determine the color of the room in the map

iex> Game.Map.room_color(%{ecology: "default"})
nil

size_of_map(zone, opts \\ [layer: 1])

size_of_map(zone :: Zone.t, layer :: integer) :: {{min_x :: integer, min_y :: integer}, {max_x :: integer, max_y :: integer}, [{{x :: integer, y :: integer}, Room.t}]}

Find the coordinates for each room in a zone and the size of the zone

1,1 is top left