ex_venture v0.15.0 Game.NPC.Actions

NPC Actions

Functions

currency_amount_to_drop(currency, rand \\ :rand)

currency_amount_to_drop(item :: Item.t, rand :: atom) :: integer

Determine how much of the currency should be dropped

Uses :rand by default

died(map)

died(state :: map) :: :ok

The NPC died, send out messages

drop_currency(room_id, npc, currency)

drop_currency(room_id :: integer, npc :: NPC.t, currency :: integer) :: :ok

Drop any currency into the room

Only when above 0

drop_item?(item, rand \\ :rand)

drop_item?(item :: Item.t, rand :: atom) :: boolean

Determine if the item should be dropped

Uses :rand by default

drop_items(room_id, npc, item_ids)

drop_items(room_id :: integer, npc :: NPC.t, item_ids :: [integer]) :: :ok

Drop items into the room with a random chance

maybe_died(stats, state)

maybe_died(stats :: map, state :: map) :: :ok

Check if the NPC died, and if so perform actions

tick(state, time)

tick(state :: map, time :: DateTime.t) :: :ok | {:update, map}

Respawn the NPC as a tick happens