ex_venture v0.15.0 Game.Session.Login

Login workflow

Displays the MOTD and asks for a login, then password. Will push off to creating an account if that is asked for.

Functions

login(user, session, socket, state)

login(user :: map, session :: pid, socket :: pid, state :: map) :: map

Sign a user in

Edit the state to be signed in and active

process(password, session, state)

sign_in(user_id, session, state)

start(socket)

start(socket :: pid) :: :ok

Start text for logging in