ex_venture v0.15.0 Networking.Protocol

Ranch protocol implementation, socket genserver

Functions

disconnect(socket)

disconnect(socket :: pid) :: :ok

Disconnect the socket

Will terminate the socket and the session

echo(socket, message)

echo(socket :: pid, message :: String.t) :: :ok

Echo a line to the client

This includes a new line at the end of the message

handle_gmcp(arg1, state)

Handle GMCP requests

Handles the following options:

Core.Supports.Set

handle_options(data, socket, fun)

Handle telnet options

Multiple IAC dos might come in at the same time, so forward them along to us after handling one.

init(ref, socket, transport)

prompt(socket, message)

prompt(socket :: pid, message :: String.t) :: :ok

Echo a prompt to the client

This does not include a new line at the end of the message

push_gmcp(socket, module, data)

push_gmcp(socket :: pid, module :: String.t, data :: String.t) :: :ok

Push GMCP data to the client

tcp_option(socket, atom, bool)

tcp_option(socket :: pid, command :: atom, toggle :: boolean) :: :ok

Toggle telnet options