ex_venture v0.15.0 Web.Router.Helpers

Module with named helpers generated from Web.Router.

Functions

bug_path(conn_or_endpoint, action)

bug_path(conn_or_endpoint, action, params)

bug_path(conn_or_endpoint, action, id, params)

bug_url(conn_or_endpoint, action)

bug_url(conn_or_endpoint, action, params)

bug_url(conn_or_endpoint, action, id, params)

class_path(conn_or_endpoint, action)

class_path(conn_or_endpoint, action, params)

class_path(conn_or_endpoint, action, id, params)

class_skill_path(conn_or_endpoint, action, class_id)

class_skill_path(conn_or_endpoint, action, class_id, params)

class_skill_url(conn_or_endpoint, action, class_id)

class_skill_url(conn_or_endpoint, action, class_id, params)

class_url(conn_or_endpoint, action)

class_url(conn_or_endpoint, action, params)

class_url(conn_or_endpoint, action, id, params)

config_path(conn_or_endpoint, action)

config_path(conn_or_endpoint, action, params)

config_url(conn_or_endpoint, action)

config_url(conn_or_endpoint, action, params)

dashboard_path(conn_or_endpoint, action)

dashboard_path(conn_or_endpoint, action, params)

dashboard_url(conn_or_endpoint, action)

dashboard_url(conn_or_endpoint, action, params)

exit_path(conn_or_endpoint, action, id)

exit_path(conn_or_endpoint, action, id, params)

exit_url(conn_or_endpoint, action, id)

exit_url(conn_or_endpoint, action, id, params)

help_topic_path(conn_or_endpoint, action)

help_topic_path(conn_or_endpoint, action, params)

help_topic_path(conn_or_endpoint, action, id, params)

help_topic_url(conn_or_endpoint, action)

help_topic_url(conn_or_endpoint, action, params)

help_topic_url(conn_or_endpoint, action, id, params)

insight_path(conn_or_endpoint, action)

insight_path(conn_or_endpoint, action, params)

insight_url(conn_or_endpoint, action)

insight_url(conn_or_endpoint, action, params)

item_path(conn_or_endpoint, action)

item_path(conn_or_endpoint, action, params)

item_path(conn_or_endpoint, action, id, params)

item_tag_path(conn_or_endpoint, action)

item_tag_path(conn_or_endpoint, action, params)

item_tag_path(conn_or_endpoint, action, id, params)

item_tag_url(conn_or_endpoint, action)

item_tag_url(conn_or_endpoint, action, params)

item_tag_url(conn_or_endpoint, action, id, params)

item_tagging_path(conn_or_endpoint, action, item_id)

item_tagging_path(conn_or_endpoint, action, item_id, params)

item_tagging_url(conn_or_endpoint, action, item_id)

item_tagging_url(conn_or_endpoint, action, item_id, params)

item_url(conn_or_endpoint, action)

item_url(conn_or_endpoint, action, params)

item_url(conn_or_endpoint, action, id, params)

npc_item_path(conn_or_endpoint, action, npc_id)

npc_item_path(conn_or_endpoint, action, npc_id, params)

npc_item_path(conn_or_endpoint, action, npc_id, id, params)

npc_item_url(conn_or_endpoint, action, npc_id)

npc_item_url(conn_or_endpoint, action, npc_id, params)

npc_item_url(conn_or_endpoint, action, npc_id, id, params)

npc_path(conn_or_endpoint, action)

npc_path(conn_or_endpoint, action, params)

npc_path(conn_or_endpoint, action, id, params)

npc_spawner_path(conn_or_endpoint, action, npc_id)

npc_spawner_path(conn_or_endpoint, action, npc_id, params)

npc_spawner_url(conn_or_endpoint, action, npc_id)

npc_spawner_url(conn_or_endpoint, action, npc_id, params)

npc_url(conn_or_endpoint, action)

npc_url(conn_or_endpoint, action, params)

npc_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

public_account_path(conn_or_endpoint, action)

public_account_path(conn_or_endpoint, action, params)

public_account_url(conn_or_endpoint, action)

public_account_url(conn_or_endpoint, action, params)

public_class_path(conn_or_endpoint, action)

public_class_path(conn_or_endpoint, action, params)

public_class_path(conn_or_endpoint, action, id, params)

public_class_url(conn_or_endpoint, action)

public_class_url(conn_or_endpoint, action, params)

public_class_url(conn_or_endpoint, action, id, params)

public_help_path(conn_or_endpoint, action)

public_help_path(conn_or_endpoint, action, params)

public_help_path(conn_or_endpoint, action, command, params)

public_help_url(conn_or_endpoint, action)

public_help_url(conn_or_endpoint, action, params)

public_help_url(conn_or_endpoint, action, command, params)

public_page_path(conn_or_endpoint, action)

public_page_path(conn_or_endpoint, action, params)

public_page_url(conn_or_endpoint, action)

public_page_url(conn_or_endpoint, action, params)

public_play_path(conn_or_endpoint, action)

public_play_path(conn_or_endpoint, action, params)

public_play_url(conn_or_endpoint, action)

public_play_url(conn_or_endpoint, action, params)

public_race_path(conn_or_endpoint, action)

public_race_path(conn_or_endpoint, action, params)

public_race_path(conn_or_endpoint, action, id, params)

public_race_url(conn_or_endpoint, action)

public_race_url(conn_or_endpoint, action, params)

public_race_url(conn_or_endpoint, action, id, params)

public_session_path(conn_or_endpoint, action)

public_session_path(conn_or_endpoint, action, params)

public_session_url(conn_or_endpoint, action)

public_session_url(conn_or_endpoint, action, params)

race_path(conn_or_endpoint, action)

race_path(conn_or_endpoint, action, params)

race_path(conn_or_endpoint, action, id, params)

race_url(conn_or_endpoint, action)

race_url(conn_or_endpoint, action, params)

race_url(conn_or_endpoint, action, id, params)

room_exit_path(conn_or_endpoint, action, room_id)

room_exit_path(conn_or_endpoint, action, room_id, params)

room_exit_url(conn_or_endpoint, action, room_id)

room_exit_url(conn_or_endpoint, action, room_id, params)

room_item_path(conn_or_endpoint, action, id)

room_item_path(conn_or_endpoint, action, id, params)

room_item_url(conn_or_endpoint, action, id)

room_item_url(conn_or_endpoint, action, id, params)

room_path(conn_or_endpoint, action, id)

room_path(conn_or_endpoint, action, id, params)

room_room_item_path(conn_or_endpoint, action, room_id)

room_room_item_path(conn_or_endpoint, action, room_id, params)

room_room_item_url(conn_or_endpoint, action, room_id)

room_room_item_url(conn_or_endpoint, action, room_id, params)

room_shop_path(conn_or_endpoint, action, room_id)

room_shop_path(conn_or_endpoint, action, room_id, params)

room_shop_url(conn_or_endpoint, action, room_id)

room_shop_url(conn_or_endpoint, action, room_id, params)

room_url(conn_or_endpoint, action, id)

room_url(conn_or_endpoint, action, id, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

shop_item_path(conn_or_endpoint, action, id)

shop_item_path(conn_or_endpoint, action, id, params)

shop_item_url(conn_or_endpoint, action, id)

shop_item_url(conn_or_endpoint, action, id, params)

shop_path(conn_or_endpoint, action, id)

shop_path(conn_or_endpoint, action, id, params)

shop_shop_item_path(conn_or_endpoint, action, shop_id)

shop_shop_item_path(conn_or_endpoint, action, shop_id, params)

shop_shop_item_url(conn_or_endpoint, action, shop_id)

shop_shop_item_url(conn_or_endpoint, action, shop_id, params)

shop_url(conn_or_endpoint, action, id)

shop_url(conn_or_endpoint, action, id, params)

skill_path(conn_or_endpoint, action, id)

skill_path(conn_or_endpoint, action, id, params)

skill_url(conn_or_endpoint, action, id)

skill_url(conn_or_endpoint, action, id, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_path(conn_or_endpoint, action, id, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action, id, params)

zone_path(conn_or_endpoint, action)

zone_path(conn_or_endpoint, action, params)

zone_path(conn_or_endpoint, action, id, params)

zone_room_path(conn_or_endpoint, action, zone_id)

zone_room_path(conn_or_endpoint, action, zone_id, params)

zone_room_url(conn_or_endpoint, action, zone_id)

zone_room_url(conn_or_endpoint, action, zone_id, params)

zone_url(conn_or_endpoint, action)

zone_url(conn_or_endpoint, action, params)

zone_url(conn_or_endpoint, action, id, params)