ex_venture v0.15.0 Web.User

Bounded context for the Phoenix app talking to the data layer

Functions

all(opts \\ [])

all(opts :: Keyword.t) :: [Data.User.t]

Load all users

change_password(user, current_password, params)

change_password(user :: Data.User.t, current_password :: String.t, params :: map) :: {:ok, Data.User.t}

Change a user’s password

connected_players()

connected_players() :: [Data.User.t]

List out connected players

disconnect()

disconnect() :: :ok

Disconnect players

The server will shutdown shortly.

from_token(token)

from_token(token :: String.t) :: Data.User.t

Fetch a user from a web token

get(id)

get(id :: integer) :: Data.User.t

Load a user

teleport(user, room_id)

teleport(user :: Data.User.t, room_id :: integer) ::
  {:ok, Data.User.t} |
  {:error, map}

Teleport a user to the room

Updates the save and sends a message to their session

teleport_player_in_game(user, room_id)